editoast: move TrackSectionCache to infra_cache
