+++ editoast/src/infra_cache/object_cache.rs
@@ -5,12 +5,14 @@ mod operational_point_cache;
 mod signal_cache;
 mod switch_cache;
 mod switch_type_cache;
+mod track_section_cache;
 
 pub use buffer_stop_cache::BufferStopCache;
 pub use detector_cache::DetectorCache;
 pub use operational_point_cache::OperationalPointCache;
 pub use signal_cache::SignalCache;
 pub use switch_cache::SwitchCache;
+pub use track_section_cache::TrackSectionCache;
 
 cfg_if! {
     if #[cfg(test)] {
